V2.0 - final updates
Fully documented the project.
Added crop that would appeared when died because of the weather/water level in soil.
Player can lose when cannot manage the water level.
Fix variables' name to provide a more meaningful message.
Optimize the min/max water level in the farm so player would not lose easily.